RHW-7 Move rolup to the root
