+++ .eslintrc.json
@@ -117,7 +117,7 @@
     "import/no-extraneous-dependencies": [
       "error",
       {
-        "devDependencies": ["**/*.test.js", "packages/*/*.js", "docs/*.js"]
+        "devDependencies": ["**/*.test.js", "packages/*/*.js", "./*.js"]
       }
     ]
   },
