+++ package.json
@@ -7,7 +7,9 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "yarn workspace @react-hook-window/use-windowed-list build",
+    "clean": "yarn workspaces run clean",
+    "prebuild": "yarn run clean",
+    "build": "rollup --config",
     "test": "jest ./tests",
     "test:watch": "yarn test -- --watch",
     "cy:open": "yarn workspace e2e cy:open",
@@ -54,6 +56,11 @@
     "jest": "^27.0.6",
     "lint-staged": "^11.0.0",
     "prettier": "^2.3.2",
+    "rollup": "^2.52.7",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-replace": "^2.2.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-typescript2": "^0.30.0",
     "typescript": "^4.3.5"
   }
 }
