+++ packages/use-windowed-list/package.json
@@ -19,8 +19,6 @@
   },
   "scripts": {
     "clean": "rm -rf ./dist",
-    "prebuild": "yarn run clean",
-    "build": "rollup --config",
     "postpublish": "PACKAGE_VERSION=$(cat package.json | grep 'version' | head -1 | awk -F: '{ print $2 }' | sed 's/[\",]//g' | tr -d '[[:space:]]') && git tag v$PACKAGE_VERSION && git push --tags",
     "typecheck": "tsc -p ./tsconfig.json --noEmit"
   },
@@ -33,14 +31,8 @@
     "normalize-scroll-left": "^0.2.1"
   },
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^13.0.0",
     "@types/lodash.debounce": "^4.0.6",
     "@types/lodash.throttle": "^4.1.6",
-    "@types/react": "^17.0.13",
-    "rollup": "^2.52.7",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0"
+    "@types/react": "^17.0.13"
   }
 }
