+++ sandbox/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@react-hook-window/use-windowed-list": "*",
+    "@react-hook-window/use-infinite-loader": "*",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.2.0",
