+++ sandbox/src/InfiniteLoaderDemo.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react'
-import {
-  useWindowedList,
-  useInfiniteLoader
-} from '@react-hook-window/use-windowed-list'
+import { useWindowedList } from '@react-hook-window/use-windowed-list'
+import { useInfiniteLoader } from '@react-hook-window/use-infinite-loader'
 
 const loadRange = (
   start: number,
